pyproject.toml
[tool.poetry]
name = "fastapi-jwt"
version = "0.1.0"
description = ""
authors = ["studentq <studentq.work@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pyjwt = "^2.8.0"
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.30"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
